requirements.txt
# ──────────────────────────────────────────────
# JoyVet Care — Python dependencies
# ──────────────────────────────────────────────

# Core framework
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-extensions==3.2.3
python-decouple==3.8

# Database
psycopg2-binary==2.9.10

# Cache / Message broker
redis==5.2.0

# Task queue
celery[redis]==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# ASGI / Real-time
channels==4.1.0
channels-redis==4.2.1
daphne==4.1.2

# Auth
djangorestframework-simplejwt==5.3.1

# Images & QR codes
Pillow==11.0.0
qrcode[pil]==8.0

# Analytics / Reporting
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.3
plotly==5.24.1

# Static files serving
whitenoise==6.8.2

# WhatsApp / SMS (optional — system works without it)
twilio==9.3.6

# Production WSGI/ASGI server
gunicorn==23.0.0

# ──────────────────────────────────────────────
# Development & Testing
# ──────────────────────────────────────────────
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
Faker==33.1.0
